utils/dnseum.py
# Placeholder for dnseum tool implementation
from telegram import Update
from telegram.ext import CallbackContext

def dnseum_query(update: Update, context: CallbackContext) -> None:
    query = ' '.join(context.args)
    if not query:
        update.message.reply_text('Veuillez fournir une requête pour dnseum.')
        return

    # Implementation of dnseum tool goes here
    update.message.reply_text(f"Résultats de dnseum pour '{query}': Fonctionnalité non encore implémentée.")
